feat: Add simple confirmation for git model deletion
Somehow a simple window.confirm() is so straightforward, that it does
not justify a separate dialog in angular.

+++ frontend/src/app/projects/models/add-git-source/add-git-source.component.ts
@@ -291,6 +291,10 @@ export class AddGitSourceComponent implements OnInit, OnDestroy {
   }
 
   deleteGitModel(): void {
+    if (!window.confirm(`Do you really want to delete this Git model?`)) {
+      return;
+    }
+
     this.gitModelService
       .deleteGitSource(
         this.projectService.project?.slug!,
